feat(fieldsearchbarfull): convert to typescript
